do brushes too
