+++ src/brushes/CircleBrush.ts
@@ -6,13 +6,7 @@ import { Group } from '../shapes/Group';
 import { getRandomInt } from '../util/internals';
 import type { Canvas } from '../canvas/Canvas';
 import { BaseBrush } from './BaseBrush';
-
-export type CircleBrushPoint = {
-  x: number;
-  y: number;
-  radius: number;
-  fill: string;
-};
+import type { CircleBrushPoint } from './typedefs';
 
 export class CircleBrush extends BaseBrush {
   /**
